Update Project Structure - Add Router
