+++ .flake8
@@ -2,6 +2,7 @@
 exclude = .git,__pycache__,venv,migrations,__init__.py
 max-complexity = 12
 import-order-style = google
+application-import-names = handlers
 max-line-length = 120
 black-config = pyproject.toml
 inline-quotes = "
