+++ bot/__main__.py
@@ -1,0 +1,31 @@
+import asyncio
+
+from aiogram import Bot, Dispatcher
+from aiogram.client.default import DefaultBotProperties
+from aiogram.enums import ParseMode
+from aiogram.fsm.storage.memory import MemoryStorage
+
+from handlers import user_handlers
+
+API_TOKEN = "your-bot-token"
+
+
+async def main() -> None:
+    # Инициализируем объект хранилища
+    storage = MemoryStorage()
+
+    bot = Bot(token=API_TOKEN, default=DefaultBotProperties(parse_mode=ParseMode.HTML))
+    dp = Dispatcher(storage=storage)
+
+    # Регистриуем роутеры
+    dp.include_router(user_handlers.router)
+
+    await bot.delete_webhook(drop_pending_updates=True)
+    await dp.start_polling(bot)
+
+
+if __name__ == "__main__":
+    asyncio.run(main())
+
+
+__all__ = []
