+++ bot/handlers/__init__.py
@@ -1,0 +1,3 @@
+# handlers
+
+from . import user as user_handlers
