--- main.py
@@ -1,33 +1,0 @@
-import asyncio
-
-from aiogram import Bot, Dispatcher, html
-from aiogram.client.default import DefaultBotProperties
-from aiogram.enums import ParseMode
-from aiogram.filters import CommandStart
-from aiogram.types import Message
-
-API_TOKEN = "your-bot-token"
-
-bot = Bot(token=API_TOKEN)
-dp = Dispatcher()
-
-
-@dp.message(CommandStart())
-async def command_start_handler(message: Message) -> None:
-    """This handler receives messages with `/start` command
-
-    Args:
-        message (Message):
-    """
-
-    await message.answer(f"Hello, {html.bold(message.from_user.full_name)}!")
-
-
-async def main() -> None:
-    bot = Bot(token=API_TOKEN, default=DefaultBotProperties(parse_mode=ParseMode.HTML))
-
-    await dp.start_polling(bot)
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
